requirements.txt
Augmentor==0.2.3
Cython==0.28.5
h5py==2.8.0
kaggle==1.4.7.1
Keras==2.2.2
Keras-Applications==1.0.4
git+https://www.github.com/farizrahman4u/keras-contrib.git #keras-contrib==2.0.8
Keras-Preprocessing==1.0.2
keras-resnet==0.1.0
matplotlib==3.0.0
numpy==1.15.1
opencv-python==3.4.3.18
pandas==0.23.4
Pillow==5.2.0
protobuf==3.6.1
pycocotools==2.0.0
pydicom==1.2.0
pydot==1.2.4
Pygments==2.2.0
pylint==2.1.1
requests==2.19.1
scikit-image==0.14.0
scikit-learn==0.19.2
scipy==1.1.0
tensorboard==1.10.0
tensorflow-gpu==1.10.1
termcolor==1.1.0
Sphinx==1.8.1
sphinx-rtd-theme==0.4.2
sphinxcontrib-websupport==1.1.0
